Revert "Make this beast of burden finally compile"
This reverts commit e491853ef404a5660e3a1d5956c206f0b9e8b1f9.
