+++ src/validation.cpp
@@ -50,7 +50,6 @@
 #include <boost/math/distributions/poisson.hpp>
 #include <boost/thread.hpp>
 #include "komodo_validation012.h"
-#include "wallet/wallet.h"
 
 using namespace std;
